+++ Anubis.LC.ExtraDays/ExtraDaysToDeadlinePlugin.cs
@@ -4,29 +4,46 @@ using LethalAPI.LibTerminal.Models;
 using Anubis.LC.ExtraDays.Commands;
 using Anubis.LC.ExtraDays.Helpers;
 using LethalAPI.LibTerminal;
+using RuntimeNetcodeRPCValidator;
+using Anubis.LC.ExtraDays.ModNetwork;
 
 namespace Anubis.LC.ExtraDays
 {
     [BepInPlugin(modGUID, modName, modVersion)]
+    [BepInDependency(MyPluginInfo.PLUGIN_GUID, MyPluginInfo.PLUGIN_VERSION)]
     [BepInDependency("ainavt.lc.lethalconfig")]
     public class ExtraDaysToDeadlinePlugin : BaseUnityPlugin
     {
-        private const string modGUID = ExtraDaysToDeadlineStaticHelper.modGUID;
-        private const string modName = ExtraDaysToDeadlineStaticHelper.modName;
-        private const string modVersion = ExtraDaysToDeadlineStaticHelper.modVersion;
+        private const string modGUID = ModStaticHelper.modGUID;
+        private const string modName = ModStaticHelper.modName;
+        private const string modVersion = ModStaticHelper.modVersion;
 
         private Harmony m_Harmony = new Harmony(modGUID);
 
         private TerminalModRegistry m_Registry;
 
+        public static ExtraDaysToDeadlinePlugin Instance;
+
+        private NetcodeValidator netcodeValidator;
+
         private void Awake()
         {
-            ExtraDaysToDeadlineStaticHelper.Logger.LogInfo($"{modGUID} is loading...");
+            if(Instance == null)
+            {
+                Instance = this;
+            }
 
-            ExtraDaysToDeadlineStaticHelper.Logger.LogInfo($"Installing patches");
+            netcodeValidator = new NetcodeValidator(ModStaticHelper.modGUID);
+            netcodeValidator.PatchAll();
+
+            netcodeValidator.BindToPreExistingObjectByBehaviour<Networking, Terminal>();
+
+            ModStaticHelper.Logger.LogInfo($"{modGUID} is loading...");
+
+            ModStaticHelper.Logger.LogInfo($"Installing patches");
             m_Harmony.PatchAll(typeof(ExtraDaysToDeadlinePlugin).Assembly);
 
-            ExtraDaysToDeadlineStaticHelper.Logger.LogInfo($"Registering built-in Commands");
+            ModStaticHelper.Logger.LogInfo($"Registering built-in Commands");
 
             // Create registry for the Terminals API
             m_Registry = TerminalRegistry.CreateTerminalRegistry();
@@ -38,7 +55,7 @@ namespace Anubis.LC.ExtraDays
 
             DontDestroyOnLoad(this);
 
-            ExtraDaysToDeadlineStaticHelper.Logger.LogInfo($"Plugin {modGUID} is loaded!");
+            ModStaticHelper.Logger.LogInfo($"Plugin {modGUID} is loaded!");
         }
     }
 }